review: move method to new navigation module
