+++ index.js
@@ -2,3 +2,4 @@ export { sendAnalyticEvent } from './src/analytics';
 export { APVideoPlayer } from './src/player';
 export { ZappPipesService } from './src/datasource/ZappPipesService';
 export { default as reminders } from './src/reminders';
+export { closeModalScreen } from './src/navigation';
