+++ src/navigation/index.js
@@ -1,0 +1,14 @@
+import { NativeModules, Platform } from 'react-native';
+
+export const closeModalScreen = Platform.select({
+  android: () => {
+    NativeModules.APReactNativeBridge.handleCommand('stop', {});
+  },
+  ios: () => {
+    NativeModules.ZPReactNativeBridgeListener.postEvent(
+      'dismiss_modal_view',
+      { animated: 1 },
+      () => {}
+    );
+  }
+});
