+++ src/utils/index.js
@@ -1,4 +1,4 @@
-import { Dimensions, NativeModules, Platform } from 'react-native';
+import { Dimensions } from 'react-native';
 
 export const getFrameDims = (maxWidth, ratio = 9 / 16) => {
   const width = maxWidth || Dimensions.get('window').width;
@@ -8,16 +8,3 @@ export const getFrameDims = (maxWidth, ratio = 9 / 16) => {
     height: getFrameHeight(width)
   };
 };
-
-export const closeModalScreen = Platform.select({
-  android: () => {
-    NativeModules.APReactNativeBridge.handleCommand('stop', {});
-  },
-  ios: () => {
-    NativeModules.ZPReactNativeBridgeListener.postEvent(
-      'dismiss_modal_view',
-      { animated: 1 },
-      () => {}
-    );
-  }
-});
